# Re-enable filtering content-to-import by renderability, but faster!