@lyw07 2018-04-24
Sorry will look into this as my first thing tomorrow morning! 
I noticed that there’s an error in Travis ci, but I didn’t see it before. Will also look into this more