@jamalex 2018-04-24
Oh, hmm, yeah -- doesn't seem to be related to the changes, and only happened in one tox env. Maybe we should just retry?

```
WARNING: autodoc: failed to import module 'kolibri.auth.test.test_api'; the following exception was raised:
No module named 'factory'
WARNING: autodoc: failed to import module 'kolibri.auth.test.test_permissions_classes'; the following exception was raised:
No module named 'mock'
Exception occurred:
  File "/home/travis/build/learningequality/kolibri/.tox/docs/lib/python3.5/site-packages/sphinx/domains/python.py", line 568, in run
    (modname, self.env.doc2path(modules[modname])),
AttributeError: 'DirectiveAdapter' object has no attribute 'env'
The full traceback has been saved in /tmp/sphinx-err-9lvbldiy.log, if you want to report the issue to the developers.
Please also report this if it was a user error, so that a better error message can be provided next time.
A bug report can be filed in the tracker at <https://github.com/sphinx-doc/sphinx/issues>. Thanks!
make[1]: *** [html] Error 2
make[1]: Leaving directory `/home/travis/build/learningequality/kolibri/docs'
make: *** [docs] Error 2
ERROR: InvocationError for command '/usr/bin/make docs' (exited with code 2)
```