@jamalex 2018-04-24
>significantly faster

0.9.0 -> ∞ (all the seconds)
0.9.1 -> 47s (with no renderability filtering)
0.9.2 -> 3.5s (full functionality reinstated)
...
0.9.3 -> it preempts your intentions and loads content before you know you want it